opt: SEQUOIADBMAINSTREAM-3732 命令位置参数Role在开启节点过滤时生效 --compile error and warning
git-svn-id: http://192.168.20.11/sequoiadb/branches/sequoiadb_2.8@36556 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/rtn/rtnCoordNodeCommands.cpp
@@ -1266,7 +1266,7 @@ namespace engine
                          "Failed to %s: failed to get the field [%s] from query",
                          _getCommandName(), FIELD_NAME_KEEP_DATA ) ;
 
-            if ( pSelfArgs->_boQuery.nFields() > validCount )
+            if ( (UINT32)pSelfArgs->_boQuery.nFields() > validCount )
             {
                rc = SDB_INVALIDARG ;
                PD_LOG( PDERROR, "Unknown parameters in command's args[%s]",
@@ -1784,7 +1784,7 @@ namespace engine
                       "Failed to %s: failed to get the field [%s] from query",
                       _getCommandName(), CMD_NAME_ENFORCED ) ;
 
-         if ( pSelfArgs->_boQuery.nFields() > validCount )
+         if ( (UINT32)pSelfArgs->_boQuery.nFields() > validCount )
          {
             rc = SDB_INVALIDARG ;
             PD_LOG( PDERROR, "Unknown parameters in command's args[%s]",
